Splitting up main.rs into modules
