+++ Cargo.toml
@@ -8,5 +8,5 @@ base64 = "0.22"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tracing-subscriber = { version = "0.3", features = ["json", "env-filter"] }
 tokio = { version = "1.37", features = ["full"] }
