+++ server/src/forwarding.rs
@@ -1,0 +1,151 @@
+use crate::models::TunneledRequest;
+use crate::AppState;
+use axum::extract::ws::Message;
+use axum::extract::State;
+use axum::extract::{Path, Query};
+use axum::http::{HeaderMap, Method, StatusCode};
+use axum::response::IntoResponse;
+use base64::engine::general_purpose;
+use base64::Engine;
+use std::collections::HashMap;
+use std::sync::Arc;
+use tokio::sync::oneshot;
+use tracing::{error, info};
+use uuid::Uuid;
+
+pub async fn handle_forwarding_request(
+    app_state: Arc<AppState>,
+    client_id: String,
+    method: Method,
+    headers: HeaderMap,
+    body: bytes::Bytes,
+    forward_path: String,
+    query_params: HashMap<String, String>,
+) -> impl IntoResponse {
+    info!(
+        "Forwarding request for client_id: {}, path: {}, method: {}, query_params: {:?}",
+        client_id,
+        forward_path,
+        method.as_str(),
+        query_params
+    );
+
+    if let Some(ws_sender) = app_state.active_websockets.get(&client_id) {
+        let headers_map: HashMap<String, String> = headers
+            .iter()
+            .map(|(k, v)| (k.to_string(), v.to_str().unwrap_or_default().to_string()))
+            .collect();
+
+        let request_id = Uuid::new_v4().to_string();
+        let tunneled_request = TunneledRequest {
+            id: request_id.clone(),
+            method: method.to_string(),
+            path: forward_path,
+            headers: headers_map,
+            query_params,
+            body: general_purpose::STANDARD.encode(body),
+        };
+
+        let (tx, rx) = oneshot::channel();
+        app_state.pending_responses.insert(request_id.clone(), tx);
+
+        match serde_json::to_string(&tunneled_request) {
+            Ok(json_payload) => {
+                if let Err(e) = ws_sender.send(Message::Text(json_payload)).await {
+                    error!("Failed to send request to websocket: {}", e);
+                    app_state.pending_responses.remove(&request_id);
+                    return (
+                        StatusCode::INTERNAL_SERVER_ERROR,
+                        "Failed to forward request to client",
+                    )
+                        .into_response();
+                }
+
+                match tokio::time::timeout(tokio::time::Duration::from_secs(30), rx).await {
+                    Ok(Ok(response)) => {
+                        let mut builder = axum::response::Response::builder().status(
+                            StatusCode::from_u16(response.status)
+                                .unwrap_or(StatusCode::INTERNAL_SERVER_ERROR),
+                        );
+
+                        for (key, value) in response.headers {
+                            builder = builder.header(key, value);
+                        }
+
+                        let body = response
+                            .body
+                            .and_then(|b| general_purpose::STANDARD.decode(b).ok());
+                        builder
+                            .body(axum::body::Body::from(body.unwrap_or_default()))
+                            .unwrap_or_else(|_| {
+                                (
+                                    StatusCode::INTERNAL_SERVER_ERROR,
+                                    "Failed to build response",
+                                )
+                                    .into_response()
+                            })
+                    }
+                    Ok(Err(_)) | Err(_) => {
+                        app_state.pending_responses.remove(&request_id);
+                        (StatusCode::GATEWAY_TIMEOUT, "Request to client timed out").into_response()
+                    }
+                }
+            }
+            Err(e) => {
+                error!("Failed to serialize request: {}", e);
+                app_state.pending_responses.remove(&request_id);
+                (
+                    StatusCode::INTERNAL_SERVER_ERROR,
+                    "Failed to serialize request",
+                )
+                    .into_response()
+            }
+        }
+    } else {
+        (StatusCode::NOT_FOUND, "Client not connected").into_response()
+    }
+}
+
+#[axum::debug_handler]
+pub async fn forward_handler_no_path(
+    State(app_state): State<Arc<AppState>>,
+    Path(client_id): Path<String>,
+    Query(query_params): Query<HashMap<String, String>>,
+    method: Method,
+    headers: HeaderMap,
+    body: bytes::Bytes,
+) -> impl IntoResponse {
+    let forward_path = "/".to_string();
+    handle_forwarding_request(
+        app_state,
+        client_id,
+        method,
+        headers,
+        body,
+        forward_path,
+        query_params,
+    )
+    .await
+}
+
+#[axum::debug_handler]
+pub async fn forward_handler_with_path(
+    State(app_state): State<Arc<AppState>>,
+    Path((client_id, path)): Path<(String, String)>,
+    Query(query_params): Query<HashMap<String, String>>,
+    method: Method,
+    headers: HeaderMap,
+    body: bytes::Bytes,
+) -> impl IntoResponse {
+    let forward_path = format!("/{}", path);
+    handle_forwarding_request(
+        app_state,
+        client_id,
+        method,
+        headers,
+        body,
+        forward_path,
+        query_params,
+    )
+    .await
+}
